Feat(norms): adapt ELI for announcement date
The announcement date is now also migrated into the metadata section
format. As the ELI depends on this date, it must be adapted to receive
this data.
But the new metadata section allows to define dates or years as
announcement date. As the ELI only needs the year anyway, the associated
parameter was adapted to receive the year directly. This way the Norm
can pass either the year metadatum or the year of the date metadatum.
The fallback case is still the old flat metadata.

RISDEV-1965