Feat(norms): db migration for announcement date
Usually, our decision was to not migrate any flat metadata from the
`norms` table to the metadata(sections). But for the announcement date
this is necessary to not break compatibility with the ELIs.
Therefore this migration first creates new announcement date sections
and metadata for all norms that that have an announcement date in the
`norms` table.
The follow up migration alters the `norms` table to the new schema.

RISDEV-1965